requirements.txt
pandas==2.3.1
pandas-datareader>=0.10.0
fredapi==0.5.2
openai==1.97.1
python-dotenv==1.1.1
ipykernel==6.30.0
pyyaml==6.0.2
psycopg2-binary==2.9.10
setuptools>=80.0.0
openpyxl==3.1.2